docs: document text output option
